+++ stonefish_slam/utils/__init__.py
@@ -1,5 +1,4 @@
-from .profiler import MappingProfiler
 from .fusion import ema_fusion
 from .sonar import OculusProperty
 
-__all__ = ['MappingProfiler', 'ema_fusion', 'OculusProperty']
+__all__ = ['ema_fusion', 'OculusProperty']